refactor: extract action interpretation into diff_entries module
