+++ src/delta_engine/application/rendering.py
@@ -2,82 +2,25 @@
 Diff and grid rendering for table and sync run reports.
 
 The report value types in report.py are pure data; all human-readable
-formatting lives here. The public entry points are render_report (status
-grid plus summary footer) and render_diff (per-table change blocks);
-render_grid, render_diff_block, run_summary_footer, and _action_entries
-are the building blocks they compose.
+formatting lives here. Action interpretation (what each action means) lives
+in diff_entries.py, the shared meaning layer this module and to_dict()
+consume. The public entry points are render_report (status grid plus summary
+footer) and render_diff (per-table change blocks); render_grid,
+render_diff_block, and run_summary_footer are the building blocks they compose.
 """
 
 from collections import Counter
-from collections.abc import Mapping
-from dataclasses import dataclass
-from enum import IntEnum
-import functools
-from types import MappingProxyType
 from typing import Final
 
+from delta_engine.application.diff_entries import (
+    CATEGORY_NOUN,
+    DiffCategory,
+    DiffEntry,
+    action_entries,
+)
 from delta_engine.application.ports import ReadFailed
 from delta_engine.application.report import SyncReport, TableRunReport
-from delta_engine.domain.model import Column, DataType, Decimal
-from delta_engine.domain.plan import (
-    Action,
-    ActionPlan,
-    AddColumn,
-    AlterClustering,
-    AlterColumnType,
-    CreateTable,
-    DropColumn,
-    DropForeignKey,
-    DropPrimaryKey,
-    SetColumnComment,
-    SetColumnNullability,
-    SetColumnTag,
-    SetForeignKey,
-    SetPrimaryKey,
-    SetProperty,
-    SetTableComment,
-    SetTableTag,
-    UnsetColumnTag,
-    UnsetProperty,
-    UnsetTableTag,
-)
-
-
-def _type_name(data_type: DataType) -> str:
-    """Backend-agnostic display name for a domain data type (e.g. 'String')."""
-    return type(data_type).__name__
-
-
-def _type_display(data_type: DataType) -> str:
-    """Display name including decimal parameters, so a precision widen is visible."""
-    if isinstance(data_type, Decimal):
-        return f"Decimal({data_type.precision},{data_type.scale})"
-    return _type_name(data_type)
-
-
-class DiffCategory(IntEnum):
-    """Diff line groups, in display order (enum value = order)."""
-
-    COLUMNS = 1
-    KEYS = 2
-    CLUSTERING = 3
-    PROPERTIES = 4
-    TAGS = 5
-    COMMENTS = 6
-
-
-# (singular, plural) nouns per category: the plural names the diff group heading;
-# the singular is used in the grid's humanized detail count.
-_CATEGORY_NOUN: Final[Mapping[DiffCategory, tuple[str, str]]] = MappingProxyType(
-    {
-        DiffCategory.COLUMNS: ("column", "columns"),
-        DiffCategory.KEYS: ("key", "keys"),
-        DiffCategory.CLUSTERING: ("clustering", "clustering"),
-        DiffCategory.PROPERTIES: ("property", "properties"),
-        DiffCategory.TAGS: ("tag", "tags"),
-        DiffCategory.COMMENTS: ("comment", "comments"),
-    }
-)
+from delta_engine.domain.plan import ActionPlan, CreateTable
 
 # Shown wherever a report has a readable state but no planned actions. One
 # spelling, two presentations: bare in the grid's DETAIL cell, parenthesised as
@@ -87,150 +30,6 @@ _NO_CHANGES: Final[str] = "no changes"
 _DETAIL_MAX_CHARS: Final[int] = 60
 
 _GRID_HEADERS: Final[tuple[str, str, str, str]] = ("TABLE", "STATUS", "ACTIONS", "DETAIL")
-
-
-@dataclass(frozen=True, slots=True)
-class DiffEntry:
-    """One rendered diff line: its category, +/-/~ symbol, and aligned cells."""
-
-    category: DiffCategory
-    symbol: str
-    cells: tuple[str, ...]
-
-
-def _column_add_entry(column: Column) -> DiffEntry:
-    """Build a '+' columns entry for a created column (name, type, optional NOT NULL)."""
-    cells = [column.name, _type_name(column.data_type)]
-    if not column.nullable:
-        cells.append("NOT NULL")
-    return DiffEntry(DiffCategory.COLUMNS, "+", tuple(cells))
-
-
-@functools.singledispatch
-def _action_entries(action: Action) -> tuple[DiffEntry, ...]:
-    """Render one plan action as one or more category-tagged diff entries."""
-    raise NotImplementedError(f"No diff entries for action {type(action).__name__}")
-
-
-@_action_entries.register
-def _(action: CreateTable) -> tuple[DiffEntry, ...]:
-    entries = [_column_add_entry(column) for column in action.table.columns]
-    primary_key_columns = action.table.primary_key_columns
-    if primary_key_columns:
-        entries.append(
-            DiffEntry(DiffCategory.KEYS, "+", (f"primary key ({', '.join(primary_key_columns)})",))
-        )
-    if action.table.clustered_by:
-        columns = ", ".join(action.table.clustered_by)
-        entries.append(DiffEntry(DiffCategory.CLUSTERING, "+", (f"clustering ({columns})",)))
-    return tuple(entries)
-
-
-@_action_entries.register
-def _(action: AddColumn) -> tuple[DiffEntry, ...]:
-    return (_column_add_entry(action.column),)
-
-
-@_action_entries.register
-def _(action: DropColumn) -> tuple[DiffEntry, ...]:
-    return (DiffEntry(DiffCategory.COLUMNS, "-", (action.column_name,)),)
-
-
-@_action_entries.register
-def _(action: SetColumnNullability) -> tuple[DiffEntry, ...]:
-    change = "drop NOT NULL (was NOT NULL)" if action.nullable else "set NOT NULL (was nullable)"
-    return (DiffEntry(DiffCategory.COLUMNS, "~", (action.column_name, change)),)
-
-
-@_action_entries.register
-def _(action: AlterColumnType) -> tuple[DiffEntry, ...]:
-    change = f"{_type_display(action.observed_type)} → {_type_display(action.data_type)}"
-    return (DiffEntry(DiffCategory.COLUMNS, "~", (action.column_name, change)),)
-
-
-@_action_entries.register
-def _(action: SetColumnComment) -> tuple[DiffEntry, ...]:
-    if action.comment:
-        text = f"column {action.column_name}: '{action.comment}'"
-    else:
-        text = f"column {action.column_name} comment (unset)"
-    return (DiffEntry(DiffCategory.COMMENTS, "~", (text,)),)
-
-
-@_action_entries.register
-def _(action: SetTableComment) -> tuple[DiffEntry, ...]:
-    text = f"table: '{action.comment}'" if action.comment else "table comment (unset)"
-    return (DiffEntry(DiffCategory.COMMENTS, "~", (text,)),)
-
-
-@_action_entries.register
-def _(action: SetProperty) -> tuple[DiffEntry, ...]:
-    if action.observed_value is None:
-        return (DiffEntry(DiffCategory.PROPERTIES, "+", (f"{action.name} = '{action.value}'",)),)
-    text = f"{action.name} = '{action.value}' (was '{action.observed_value}')"
-    return (DiffEntry(DiffCategory.PROPERTIES, "~", (text,)),)
-
-
-@_action_entries.register
-def _(action: UnsetProperty) -> tuple[DiffEntry, ...]:
-    return (DiffEntry(DiffCategory.PROPERTIES, "-", (action.name,)),)
-
-
-@_action_entries.register
-def _(action: SetTableTag) -> tuple[DiffEntry, ...]:
-    return (DiffEntry(DiffCategory.TAGS, "~", (f"{action.name} = '{action.value}'",)),)
-
-
-@_action_entries.register
-def _(action: UnsetTableTag) -> tuple[DiffEntry, ...]:
-    return (DiffEntry(DiffCategory.TAGS, "-", (action.name,)),)
-
-
-@_action_entries.register
-def _(action: SetColumnTag) -> tuple[DiffEntry, ...]:
-    text = f"column {action.column_name}.{action.name} = '{action.value}'"
-    return (DiffEntry(DiffCategory.TAGS, "~", (text,)),)
-
-
-@_action_entries.register
-def _(action: UnsetColumnTag) -> tuple[DiffEntry, ...]:
-    return (DiffEntry(DiffCategory.TAGS, "-", (f"column {action.column_name}.{action.name}",)),)
-
-
-@_action_entries.register
-def _(action: SetPrimaryKey) -> tuple[DiffEntry, ...]:
-    return (DiffEntry(DiffCategory.KEYS, "+", (f"primary key ({', '.join(action.columns)})",)),)
-
-
-@_action_entries.register
-def _(action: DropPrimaryKey) -> tuple[DiffEntry, ...]:
-    return (DiffEntry(DiffCategory.KEYS, "-", ("primary key",)),)
-
-
-@_action_entries.register
-def _(action: SetForeignKey) -> tuple[DiffEntry, ...]:
-    text = f"foreign key ({', '.join(action.local_columns)}) → {action.referenced_table}"
-    return (DiffEntry(DiffCategory.KEYS, "+", (text,)),)
-
-
-@_action_entries.register
-def _(action: DropForeignKey) -> tuple[DiffEntry, ...]:
-    return (DiffEntry(DiffCategory.KEYS, "-", (f"foreign key {action.constraint_name}",)),)
-
-
-_OPTIMIZE_FULL_HINT: Final[str] = "run OPTIMIZE FULL to recluster existing data"
-
-
-@_action_entries.register
-def _(action: AlterClustering) -> tuple[DiffEntry, ...]:
-    # No hint on removal: OPTIMIZE FULL errors on a table without clustering
-    # columns (DELTA_OPTIMIZE_FULL_NOT_SUPPORTED); existing files simply keep
-    # their old layout after CLUSTER BY NONE.
-    if not action.columns:
-        return (DiffEntry(DiffCategory.CLUSTERING, "-", ("clustering",)),)
-    columns = ", ".join(action.columns)
-    text = f"clustering ({columns}) — {_OPTIMIZE_FULL_HINT}"
-    return (DiffEntry(DiffCategory.CLUSTERING, "~", (text,)),)
 
 
 def _plan_creates_table(plan: ActionPlan) -> bool:
@@ -244,7 +43,7 @@ def _render_entry_groups(entries: list[DiffEntry]) -> list[str]:
         group = [entry for entry in entries if entry.category is category]
         if not group:
             continue
-        lines.append(f"  {_CATEGORY_NOUN[category][1]}")
+        lines.append(f"  {CATEGORY_NOUN[category][1]}")
         widths: dict[int, int] = {}
         for entry in group:
             for index, cell in enumerate(entry.cells):
@@ -266,7 +65,7 @@ def render_diff_block(report: TableRunReport) -> str:
         return f"{header}\n  ({_NO_CHANGES})"
     if _plan_creates_table(report.plan):
         header = f"{header}  (CREATE)"
-    entries = [entry for action in report.plan for entry in _action_entries(action)]
+    entries = [entry for action in report.plan for entry in action_entries(action)]
     return "\n".join([header, *_render_entry_groups(entries)])
 
 
@@ -285,13 +84,13 @@ def _humanized_action_summary(plan: ActionPlan) -> str:
     """Summarise a plan as per-category change counts in display order, e.g. '2 columns, 1 key'."""
     counts: Counter[DiffCategory] = Counter()
     for action in plan:
-        for entry in _action_entries(action):
+        for entry in action_entries(action):
             counts[entry.category] += 1
     parts: list[str] = []
     for category in DiffCategory:
         count = counts.get(category, 0)
         if count:
-            singular, plural = _CATEGORY_NOUN[category]
+            singular, plural = CATEGORY_NOUN[category]
             parts.append(f"{count} {singular if count == 1 else plural}")
     return ", ".join(parts) or _NO_CHANGES
